+++ model_compression_toolkit/core/pytorch/hessian/activation_trace_hessian_calculator_pytorch.py
@@ -15,14 +15,18 @@
 
 from typing import List, Tuple, Dict, Any
 
+from torch import autograd
 from tqdm import tqdm
 import numpy as np
 
+from model_compression_toolkit.constants import MIN_JACOBIANS_ITER, JACOBIANS_COMP_TOLERANCE
 from model_compression_toolkit.core.common import Graph
 from model_compression_toolkit.core.common.hessian import TraceHessianRequest, TraceHessianGranularity
 from model_compression_toolkit.core.common.hessian.trace_hessian_config import TraceHessianConfig
+from model_compression_toolkit.core.pytorch.hessian.pytorch_model_gradients import PytorchModelGradients
 from model_compression_toolkit.core.pytorch.hessian.trace_hessian_calculator_pytorch import \
     TraceHessianCalculatorPytorch
+from model_compression_toolkit.core.pytorch.utils import torch_tensor_to_numpy
 from model_compression_toolkit.logger import Logger
 import torch
 
@@ -58,231 +62,81 @@ class ActivationTraceHessianCalculatorPytorch(TraceHessianCalculatorPytorch):
             List[float]: Approximated trace of the Hessian for an interest point.
         """
         if self.hessian_request.granularity == TraceHessianGranularity.PER_TENSOR:
-            raise NotImplemented
+            # Set inputs to require_grad
+            for input_tensor in self.input_images:
+                input_tensor.requires_grad_()
 
-    #         output_list = self._get_model_output_replacement()
-    #
-    #         # Record operations for automatic differentiation
-    #         with tf.GradientTape(persistent=True, watch_accessed_variables=False) as g:
-    #             outputs, interest_points_tensors = self._get_model_outputs_for_single_image(output_list,
-    #                                                                                         gradient_tape=g)
-    #
-    #             # Unfold and concatenate all outputs to form a single tensor
-    #             unfold_outputs = []
-    #             for output in outputs:
-    #                 if isinstance(output, List):
-    #                     unfold_outputs += output
-    #                 else:
-    #                     unfold_outputs.append(output)
-    #
-    #             r_outputs = [tf.reshape(output, shape=[output.shape[0], -1]) for output in unfold_outputs]
-    #
-    #             # Ensure all outputs have the same shape for concatenation
-    #             concat_axis_dim = [o.shape[0] for o in r_outputs]
-    #             if not all(d == concat_axis_dim[0] for d in concat_axis_dim):
-    #                 Logger.critical(
-    #                     "Can't concat model's outputs for gradients calculation since the shape of the first axis " # pragma: no cover
-    #                     "is not equal in all outputs.")
-    #
-    #             output = tf.concat(r_outputs, axis=1)
-    #
-    #             # List to store the approximated trace of the Hessian for each interest point
-    #             trace_approx_by_node = []
-    #             for ipt in tqdm(interest_points_tensors):  # Per Interest point activation tensor
-    #                 trace_jv = []
-    #                 for j in range(self.hessian_config.num_iterations):  # Approximation iterations
-    #                     # Getting a random vector with normal distribution
-    #                     v = tf.random.normal(shape=output.shape)
-    #                     f_v = tf.reduce_sum(v * output)
-    #
-    #                     with g.stop_recording():
-    #                         # Computing the approximation by getting the gradient of (output * v)
-    #                         jac_v = g.gradient(f_v, ipt, unconnected_gradients=tf.UnconnectedGradients.ZERO)
-    #                         jac_v = tf.reshape(jac_v, [jac_v.shape[0], -1])
-    #                         jac_trace_approx = tf.reduce_mean(tf.reduce_sum(tf.pow(jac_v, 2.0)))
-    #
-    #                         # If the change to the mean approximation is insignificant we stop the calculation
-    #                         if j > MIN_JACOBIANS_ITER:
-    #                             new_mean = np.mean([jac_trace_approx, *trace_jv])
-    #                             delta = new_mean - np.mean(trace_jv)
-    #                             if np.abs(delta) / (np.abs(new_mean) + 1e-6) < JACOBIANS_COMP_TOLERANCE:
-    #                                 trace_jv.append(jac_trace_approx)
-    #                                 break
-    #
-    #                         trace_jv.append(jac_trace_approx)
-    #                 trace_approx_by_node.append(2 * tf.reduce_mean(trace_jv) / output.shape[
-    #                     -1])  # Get averaged squared trace approximation
-    #
-    #             trace_approx_by_node = tf.reduce_mean([trace_approx_by_node],
-    #                                                    axis=0)  # Just to get one tensor instead of list of tensors with single element
-    #
-    #             return trace_approx_by_node.numpy().tolist()
-    #     else:
-    #         Logger.error(f"{self.hessian_request.granularity} is not supported for Keras activation hessian's trace approx calculator")
-    #
-    #
-    # def _update_ips_with_outputs_replacements(self,
-    #                                           outputs_replacement_nodes: List[BaseNode],
-    #                                           interest_points: List[BaseNode]):
-    #     """
-    #     Updates the list of interest points with the set of pre-calculated replacement outputs.
-    #     Also, returns the indices of all output nodes (original, replacements and nodes in between them) in a
-    #     topological sorted interest points list (for later use in gradients computation and normalization).
-    #
-    #     Returns: A list of indices of the output nodes in the sorted interest points list.
-    #
-    #     """
-    #
-    #     replacement_outputs_to_ip = [r_node for r_node in outputs_replacement_nodes if
-    #                                  r_node not in interest_points]
-    #     updated_interest_points = interest_points + replacement_outputs_to_ip
-    #
-    #     # Re-sort interest points in a topological order according to the graph's sort
-    #     interest_points = [n for n in self.graph.get_topo_sorted_nodes() if n in updated_interest_points]
-    #
-    #     output_indices = [interest_points.index(n.node) for n in self.graph.get_outputs()]
-    #     replacement_indices = [interest_points.index(n) for n in outputs_replacement_nodes]
-    #     return list(set(output_indices + replacement_indices))
-    #
-    # def _get_model_output_replacement(self) -> List[str]:
-    #     """
-    #     If a model's output node is not compatible for the task of gradients computation we need to find a predecessor
-    #     node in the model's graph representation which is compatible and use it for the gradients' computation.
-    #     This method searches for this predecessor node for each output of the model.
-    #
-    #     Returns: A list of output replacement nodes.
-    #
-    #     """
-    #
-    #     replacement_outputs = []
-    #     for n in self.graph.get_outputs():
-    #         prev_node = n.node
-    #         while not self.fw_impl.is_node_compatible_for_metric_outputs(prev_node):
-    #             prev_node = self.graph.get_prev_nodes(prev_node)
-    #             assert len(prev_node) == 1, "A none compatible output node has multiple inputs, " \
-    #                                         "which is incompatible for metric computation."
-    #             prev_node = prev_node[0]
-    #         replacement_outputs.append(prev_node)
-    #     return replacement_outputs
-    #
-    # def _get_model_outputs_for_single_image(self,
-    #                                         output_list: List[str],
-    #                                         gradient_tape: tf.GradientTape) -> Tuple[List[tf.Tensor], List[tf.Tensor]]:
-    #     """
-    #     Computes the model's output according to the given graph representation on the given input,
-    #     while recording necessary intermediate tensors for gradients computation.
-    #
-    #     Args:
-    #         output_list: List of nodes that considered as model's output for the purpose of gradients computation.
-    #         gradient_tape: A GradientTape object for recording necessary info for computing gradients.
-    #
-    #     Returns: A list of output tensors and a list of activation tensors of all interest points.
-    #
-    #     """
-    #     model_input_tensors = {inode: self.fw_impl.to_tensor(self.input_images[i]) for i, inode in
-    #                            enumerate(self.graph.get_inputs())}
-    #
-    #     node_to_output_tensors_dict = dict()
-    #
-    #     # Build an OperationHandler to handle conversions from graph nodes to Keras operators.
-    #     oh = OperationHandler(self.graph)
-    #     input_nodes_to_input_tensors = {inode: tf.convert_to_tensor(model_input_tensors[inode]) for
-    #                                     inode in self.graph.get_inputs()}  # Cast numpy array to tf.Tensor
-    #
-    #     interest_points_tensors = []
-    #     output_tensors = []
-    #     for n in oh.node_sort:
-    #         # Build a dictionary from node to its output tensors, by applying the layers sequentially.
-    #         op_func = oh.get_node_op_function(n)  # Get node operation function
-    #
-    #         input_tensors = self._build_input_tensors_list(n,
-    #                                                        self.graph,
-    #                                                        node_to_output_tensors_dict)  # Fetch Node inputs
-    #
-    #         out_tensors_of_n = self._run_operation(n,  # Run node operation and fetch outputs
-    #                                                input_tensors,
-    #                                                op_func,
-    #                                                input_nodes_to_input_tensors)
-    #
-    #         # Gradients can be computed only on float32 tensors
-    #         if isinstance(out_tensors_of_n, list):
-    #             for i, t in enumerate(out_tensors_of_n):
-    #                 out_tensors_of_n[i] = tf.dtypes.cast(t, tf.float32)
-    #         else:
-    #             out_tensors_of_n = tf.dtypes.cast(out_tensors_of_n, tf.float32)
-    #
-    #         if n.name==self.hessian_request.target_node.name:
-    #             # Recording the relevant feature maps onto the gradient tape
-    #             gradient_tape.watch(out_tensors_of_n)
-    #             interest_points_tensors.append(out_tensors_of_n)
-    #         if n in output_list:
-    #             output_tensors.append(out_tensors_of_n)
-    #
-    #         if isinstance(out_tensors_of_n, list):
-    #             node_to_output_tensors_dict.update({n: out_tensors_of_n})
-    #         else:
-    #             node_to_output_tensors_dict.update({n: [out_tensors_of_n]})
-    #
-    #     return output_tensors, interest_points_tensors
-    #
-    # def _build_input_tensors_list(self,
-    #                               node: BaseNode,
-    #                               graph: Graph,
-    #                               node_to_output_tensors_dict: Dict[BaseNode, List[TFReference]]) -> List[List[TFReference]]:
-    #     """
-    #     Given a node, build a list of input tensors the node gets. The list is built
-    #     based on the node's incoming edges and previous nodes' output tensors.
-    #
-    #     Args:
-    #         node: Node to build its input tensors list.
-    #         graph: Graph the node is in.
-    #         node_to_output_tensors_dict: A dictionary from a node to its output tensors.
-    #
-    #     Returns:
-    #         A list of the node's input tensors.
-    #     """
-    #
-    #     input_tensors = []
-    #     # Go over a sorted list of the node's incoming edges, and for each source node get its output tensors.
-    #     # Append them in a result list.
-    #     for ie in graph.incoming_edges(node, sort_by_attr=EDGE_SINK_INDEX):
-    #         _input_tensors = [node_to_output_tensors_dict[ie.source_node][ie.source_index]]
-    #         input_tensors.append(_input_tensors)
-    #     return input_tensors
-    #
-    # def _run_operation(self,
-    #                    n: BaseNode,
-    #                    input_tensors: List[List[TFReference]],
-    #                    op_func: Layer,
-    #                    input_nodes_to_input_tensors: Dict[BaseNode, Any]) -> List[TFReference]:
-    #     """
-    #     Applying the layer (op_func) to the input tensors (input_tensors).
-    #
-    #     Args:
-    #         n: The corresponding node of the layer it runs.
-    #         input_tensors: List of references to Keras tensors that are the layer's inputs.
-    #         op_func: Layer to apply to the input tensors.
-    #         input_nodes_to_input_tensors: A dictionary from a node to its input tensors.
-    #
-    #     Returns:
-    #         A list of references to Keras tensors. The layer's output tensors after applying the
-    #         layer to the input tensors.
-    #     """
-    #
-    #     if len(input_tensors) == 0:  # Placeholder handling
-    #         out_tensors_of_n = input_nodes_to_input_tensors[n]
-    #     else:
-    #         input_tensors = [tensor for tensor_list in input_tensors for tensor in tensor_list]  # flat list of lists
-    #         # Build a functional node using its args
-    #         if isinstance(n, FunctionalNode):
-    #             if n.inputs_as_list:  # If the first argument should be a list of tensors:
-    #                 out_tensors_of_n = op_func(input_tensors, *n.op_call_args, **n.op_call_kwargs)
-    #             else:  # If the input tensors should not be a list but iterated:
-    #                 out_tensors_of_n = op_func(*input_tensors, *n.op_call_args, **n.op_call_kwargs)
-    #         else:
-    #             # If operator expects a single input tensor, it cannot be a list as it should have a dtype field.
-    #             if len(input_tensors) == 1:
-    #                 input_tensors = input_tensors[0]
-    #             out_tensors_of_n = op_func(input_tensors)
-    #
-    #     return out_tensors_of_n
+            model_grads_net = PytorchModelGradients(graph_float=self.graph,
+                                                    trace_hessian_request=self.hessian_request
+                                                    # interest_points=interest_points,
+                                                    # output_list=output_list
+                                                    )
+
+            # Run model inference
+            output_tensors = model_grads_net(self.input_images)
+            device = output_tensors[0].device
+
+            # Concat outputs
+            # First, we need to unfold all outputs that are given as list, to extract the actual output tensors
+            unfold_outputs = []
+            for output in output_tensors:
+                if isinstance(output, List):
+                    unfold_outputs += output
+                else:
+                    unfold_outputs.append(output)
+
+            r_outputs = [torch.reshape(output, shape=[output.shape[0], -1]) for output in unfold_outputs]
+
+            concat_axis_dim = [o.shape[0] for o in r_outputs]
+            if not all(d == concat_axis_dim[0] for d in concat_axis_dim):
+                Logger.critical(
+                    "Can't concat model's outputs for gradients calculation since the shape of the first axis "  # pragma: no cover
+                    "is not equal in all outputs.")
+
+            output = torch.concat(r_outputs, dim=1)
+
+            ipts_jac_trace_approx = []
+            for ipt in tqdm(model_grads_net.interest_points_tensors):  # Per Interest point activation tensor
+                trace_jv = []
+                for j in range(self.hessian_config.num_iterations):  # Approximation iterations
+                    # Getting a random vector with normal distribution
+                    v = torch.randn(output.shape, device=device)
+                    f_v = torch.sum(v * output)
+
+                    # Computing the jacobian approximation by getting the gradient of (output * v)
+                    jac_v = autograd.grad(outputs=f_v,
+                                          inputs=ipt,
+                                          retain_graph=True,
+                                          allow_unused=True)[0]
+                    if jac_v is None:
+                        # In case we have an output node, which is an interest point, but it is not differentiable,
+                        # we still want to set some weight for it. For this, we need to add this dummy tensor to the ipt
+                        # jacobian traces list.
+                        trace_jv.append(torch.tensor([0.0],
+                                                     requires_grad=True,
+                                                     device=device))
+                        break
+                    jac_v = torch.reshape(jac_v, [jac_v.shape[0], -1])
+                    jac_trace_approx = torch.mean(torch.sum(torch.pow(jac_v, 2.0)))
+
+                    # If the change to the mean Jacobian approximation is insignificant we stop the calculation
+                    if j > MIN_JACOBIANS_ITER:
+                        new_mean = torch.mean(torch.stack([jac_trace_approx, *trace_jv]))
+                        delta = new_mean - torch.mean(torch.stack(trace_jv))
+                        if torch.abs(delta) / (torch.abs(new_mean) + 1e-6) < JACOBIANS_COMP_TOLERANCE:
+                            trace_jv.append(jac_trace_approx)
+                            break
+
+                    trace_jv.append(jac_trace_approx)
+                ipts_jac_trace_approx.append(2 * torch.mean(torch.stack(trace_jv)) / output.shape[
+                    -1])  # Get averaged jacobian trace approximation
+
+            ipts_jac_trace_approx = torch_tensor_to_numpy(torch.Tensor(
+                ipts_jac_trace_approx))  # Just to get one tensor instead of list of tensors with single element
+
+
+            return ipts_jac_trace_approx.tolist()
+            # if norm_weights:
+            #     return _normalize_weights(ipts_jac_trace_approx, all_outputs_indices, alpha)
+            # else:
+            #     return ipts_jac_trace_approx
