+++ model_compression_toolkit/qat/pytorch/quantization_facade.py
@@ -149,14 +149,15 @@ if FOUND_TORCH:
 
         fw_impl = PytorchImplementation()
 
-        tg, bit_widths_config = core_runner(in_model=in_model,
-                                            representative_data_gen=representative_data_gen,
-                                            core_config=core_config,
-                                            fw_info=DEFAULT_PYTORCH_INFO,
-                                            fw_impl=fw_impl,
-                                            tpc=target_platform_capabilities,
-                                            target_kpi=target_kpi,
-                                            tb_w=tb_w)
+        # Ignore trace hessian service as we do not use it here
+        tg, bit_widths_config, _ = core_runner(in_model=in_model,
+                                               representative_data_gen=representative_data_gen,
+                                               core_config=core_config,
+                                               fw_info=DEFAULT_PYTORCH_INFO,
+                                               fw_impl=fw_impl,
+                                               tpc=target_platform_capabilities,
+                                               target_kpi=target_kpi,
+                                               tb_w=tb_w)
 
         tg = ptq_runner(tg, representative_data_gen, core_config, fw_info, fw_impl, tb_w)
